backend/requirements.txt
asn1crypto==0.24.0
attrdict==2.0.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coincurve==9.0.0
cytoolz==0.9.0.1
Django==2.1.2
django-web3-auth==0.1.0
eth-abi==1.2.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==1.3.0
eth-utils==1.2.2
ethereum==2.3.2
future==0.16.0
hexbytes==0.1.0
idna==2.7
lru-dict==1.1.6
parsimonious==0.8.0
pbkdf2==1.3
py-ecc==1.4.3
pycparser==2.19
pycryptodome==3.6.6
pyethash==0.1.27
pysha3==1.0.2
pytz==2018.5
PyYAML==3.13
repoze.lru==0.7
requests==2.19.1
rlp==1.0.1
scrypt==0.8.6
six==1.11.0
toolz==0.9.0
urllib3==1.23
web3==4.7.2
websockets==6.0
